+++ src/componnets/pages/Dashboard/Manage Products/ManageProducts.js
@@ -1,6 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import { toast } from 'react-toastify';
-
+import { confirmAlert } from 'react-confirm-alert';
+import 'react-confirm-alert/src/react-confirm-alert.css';
 const ManageProducts = () => {
     const [services, setServices] = useState([])
     const [Delete, setDelete] = useState(false)
@@ -9,7 +10,7 @@ const ManageProducts = () => {
             method: 'GET',
             headers: {
                 'Content-Type': 'application/json',
-                'authorization':`Bearer ${localStorage.getItem("AssesToken")}`
+                'authorization': `Bearer ${localStorage.getItem("AssesToken")}`
             },
         })
             .then(response => response.json())
@@ -18,19 +19,37 @@ const ManageProducts = () => {
             })
     }, [Delete])
     const ManageProductsDelete = id => {
-        fetch(`https://vast-ridge-73699.herokuapp.com/service/delete/${id}`, {
-            method: 'DELETE',
-            headers: {
-                'Content-Type': 'application/json',
-            },
-        })
-            .then(response => response.json())
-            .then(data => {
-                if(data){
-                    toast("Delete SuccessFully")
-                    setDelete(true)
+        confirmAlert({
+            title: 'Confirm to submit',
+            message: 'Are you sure Delete.',
+            buttons: [
+                {
+                    label: 'Yes',
+                    onClick: () => {
+                        fetch(`https://vast-ridge-73699.herokuapp.com/service/delete/${id}`, {
+                            method: 'DELETE',
+                            headers: {
+                                'Content-Type': 'application/json',
+                            },
+                        })
+                            .then(response => response.json())
+                            .then(data => {
+                                if (data) {
+                                    toast("Delete SuccessFully")
+                                    setDelete(true)
+                                }
+                            })
+                    }
+                },
+                {
+                    label: 'No',
+                    onClick: () => {
+                        toast("Note Delete")
+                    }
                 }
-            })
+            ]
+        });
+
     }
     return (
         <div>
@@ -57,7 +76,7 @@ const ManageProducts = () => {
                                     <td>{service.AvailableStok}</td>
                                     <td>${service.prices}</td>
                                     <td><button onClick={() => ManageProductsDelete(service._id)}>
-                                    <button className="btn btn-circle btn-outline hover:bg-red-600 hover:border-0">
+                                        <button className="btn btn-circle btn-outline hover:bg-red-600 hover:border-0">
                                             <svg xmlns="http://www.w3.org/2000/svg" className="h-6 w-6" fill="none" viewBox="0 0 24 24" stroke="currentColor"><path strokeLinecap="round" strokeLinejoin="round" strokeWidth="2" d="M6 18L18 6M6 6l12 12" /></svg>
                                         </button>
                                     </button></td>
